requirements.txt
alembic==1.4.2
aniso8601==8.0.0
appdirs==1.4.3
attrs==19.3.0
bcrypt==3.1.7
black==19.10b0
cffi==1.14.0
click==7.1.1
cryptography==2.9
distlib==0.3.0
dnspython==1.16.0
email-validator==1.0.5
entrypoints==0.3
fastapi==0.54.1
filelock==3.0.12
flake8==3.7.9
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
future==0.18.2
h11==0.9.0
httptools==0.1.1
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-sqlalchemy==0.22.3
mccabe==0.6.1
more-itertools==8.2.0
mypy==0.770
mypy-extensions==0.4.3
packaging==20.3
pathspec==0.7.0
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.1
pycodestyle==2.5.0
pycparser==2.20
pydantic==1.4
pyflakes==2.1.1
PyJWT==1.7.1
pyparsing==2.4.6
pyrsistent==0.16.0
pytest==5.4.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
regex==2020.4.4
six==1.14.0
SQLAlchemy==1.3.15
starlette==0.13.2
toml==0.10.0
tox==3.14.6
typed-ast==1.4.1
typing-extensions==3.7.4.2
uvicorn==0.11.3
uvloop==0.14.0
virtualenv==20.0.16
wcwidth==0.1.9
websockets==8.1
Werkzeug==1.0.1
WTForms==2.2.1
zipp==3.1.0
